// src/network/routes/emr/interfaces/patient-transfer/patient-transfer.request.ts
export interface IUpdateTransferPasien {
    id: string
    id_dokter_dpjp: string
    id_dokter_operator: string
    tanggal_transfer: string
    indikasi_transfer: string
    pengantar: string
    penerima: string
    tanggal_masuk_rs: string
    diagnosa: string
    kesadaran: string
    vital_td: string
    vital_n: string
    vital_p: string
    vital_t: string
    vital_sat_o2: string
    visus_od: string
    visus_os: string
    tonometer_od: string
    tonometer_os: string
    skala_nyeri: string
    puasa: string
    waktu_puasa: string
    keluhan: string
    alderette: string
    alderette_aktivitas: string
    alderette_sirkulasi: string
    alderette_pernafasan: string
    alderette_kesadaran: string
    alderette_warna_kulit: string
    alderette_score: string
    steward: string
    steward_kesadaran: string
    steward_pernafasan: string
    steward_motorik: string
    steward_score: string
    pemeriksaan_alat_ekg: string
    pemeriksaan_alat_laboratorium: string
    pemeriksaan_alat_laboratorium_text: string
    pemeriksaan_alat_usg: string
    pemeriksaan_alat_biometri: string
    pemeriksaan_alat_oct_macula: string
    pemeriksaan_alat_thorax_foto: string
    pemeriksaan_alat_ct_scan: string
    pemeriksaan_alat_foto_fundus: string
    pemeriksaan_alat_oct_papil: string
    pemeriksaan_alat_lain_lain: string
    pemeriksaan_alat_lain_text: string
    terapi: string
    rencana: string
    diet: string
    tanda_tangan_perawat_pengantar: string
    id_tanda_tangan_perawat_pengantar: string
    tanda_tangan_perawat_penerima: string
    id_tanda_tangan_perawat_penerima: string
    unit: string;
    emr_id: string
    nomor_mr: string
    id_pelayanan: string
    kode_cabang: string
    tipe_pasien: string
    jenis_pelayanan: string
    id_dokter: string
    no_sep: string
  }